+++ Plugin/RemoveAdminOnlyPaymentMethodFromCheckout.php
@@ -1,18 +1,18 @@
 <?php
 
 /**
- * @copyright Copyright © 2021 Orba Sp. z o.o. All rights reserved.
- * @author    info@orba.co
+ * Copyright © 2023 Lingaro sp. z o.o. All rights reserved.
+ * See LICENSE for license details.
  */
 
 declare(strict_types=1);
 
-namespace Orba\MicroFeatures\Plugin;
+namespace Lingaro\MicroFeatures\Plugin;
 
 use Magento\Payment\Model\Checks\CanUseCheckout;
 use Magento\Payment\Model\MethodInterface;
 use Magento\Quote\Model\Quote;
-use Orba\MicroFeatures\Model\Config;
+use Lingaro\MicroFeatures\Model\Config;
 
 class RemoveAdminOnlyPaymentMethodFromCheckout
 {