+++ view/frontend/requirejs-config.js
@@ -1,13 +1,13 @@
 /**
- * @copyright Copyright © 2021 Orba. All rights reserved.
- * @author    info@orba.co
+ * Copyright © 2023 Lingaro sp. z o.o. All rights reserved.
+ * See LICENSE for license details.
  */
 
 var config = {
     "config": {
         "mixins": {
             "Magento_Checkout/js/view/summary/cart-items": {
-                "Orba_MicroFeatures/js/checkout/view/summary/cart-items-mixin": true
+                "Lingaro_MicroFeatures/js/checkout/view/summary/cart-items-mixin": true
             }
         }
     }